pyproject.toml
[tool.poetry]
name = "kalokohan"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kalokohan"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^3.2"
django-extensions = "^3.2.1"
gunicorn = "^20.1.0"
dj-database-url = "^1.0.0"
whitenoise = "^6.2.0"
psycopg2 = "^2.9.3"
sentry-sdk = "^1.9.8"
factory-boy = "^3.2.1"
django-ninja = "^0.19.1"
django-solo = "^2.0.0"
requests = "^2.28.1"
responses = "^0.22.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
bandit = "^1.7.4"
detect-secrets = "^1.3.0"
django-types = "^0.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
